add enable option to files
