+++ src/index.ts
@@ -1,6 +1,6 @@
-import { addDataAndFileToRequest, type Config } from 'payload'
+import { type Config } from 'payload'
 
-import { getIdFromUrl } from './lib/getIdFromUrl.js'
+import { endpoints } from './endpoints/open-folder.js'
 
 export type PayloadFolderTreeViewConfig = {
   /**
@@ -9,6 +9,14 @@ export type PayloadFolderTreeViewConfig = {
    * @default false
    */
   disabled?: boolean
+  /**
+   * Show files in the folder tree view.
+   * This will enable the `/open-folder` endpoint to fetch files from folders.
+   * Can cause performance issues with large folder structures.
+   *
+   * @default true
+   */
+  showFiles?: boolean
 }
 
 export const payloadFolderTreeView =
@@ -16,6 +24,18 @@ export const payloadFolderTreeView =
     (config: Config): Config => {
       if (pluginOptions.disabled) {
         return config
+      }
+
+      const configEndpoints = [
+        ...(config.endpoints ?? []),
+      ]
+
+      if (pluginOptions.showFiles) {
+        const pluginEndpoints = endpoints(pluginOptions);
+
+        configEndpoints.push(
+          pluginEndpoints.openFolder,
+        )
       }
 
       return {
@@ -33,73 +53,7 @@ export const payloadFolderTreeView =
         },
         endpoints: [
           ...(config.endpoints ?? []),
-          {
-            handler: async (req) => {
-              await addDataAndFileToRequest(req);
-
-              const folderId = getIdFromUrl(req.url ?? "");
-
-              if (!folderId) {
-                return Response.json({ error: "Invalid folder ID" }, { status: 400 });
-              }
-
-              const files = await req.payload.findByID({
-                id: folderId,
-                collection: 'payload-folders',
-              })
-
-              if (!files) {
-                return Response.json({ error: "Folder not found" }, { status: 404 });
-              }
-
-
-              const mappedFilesFromFolders = [];
-
-              for (const file of files.documentsAndFolders.docs) {
-                if (file.relationTo === "payload-folders") {
-                  continue;
-                }
-                const { relationTo } = file;
-                const {
-                  createdAt: _createdAt,
-                  documentsAndFolders: _documentsAndFolders,
-                  folder: _folder,
-                  updatedAt: _updatedAt,
-                  ...rest
-                } = file.value;
-
-                mappedFilesFromFolders.push({
-                  ...rest,
-                  relationTo,
-                });
-              }
-
-              const mappedFilesFromIds = [];
-
-              for (const file of mappedFilesFromFolders) {
-                const document = await req.payload.findByID({
-                  id: file.id,
-                  collection: file.relationTo,
-                })
-
-                if (!document) {
-                  continue;
-                }
-
-                const { folder: _folder, ...rest } = document;
-
-                mappedFilesFromIds.push({
-                  ...rest,
-                  relationTo: file.relationTo,
-                });
-              }
-
-
-              return Response.json(mappedFilesFromIds)
-            },
-            method: 'get',
-            path: '/:id/folder-tree-view/open-folder',
-          }
+          ...configEndpoints,
         ],
       };
     }
