+++ src/types.ts
@@ -1,0 +1,5 @@
+import type { Endpoint } from "payload";
+
+export interface Endpoints {
+  openFolder: Omit<Endpoint, 'root'>;
+}
